dcae-poc/pyproject.toml
[tool.poetry]
name = "dcae-poc"
version = "0.1.0"
description = "DCAE - Disciplined Consensus-Driven Agentic Engineering POC"
authors = ["Sheldon <sheldon@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.5"
pydantic-settings = "^2.1"
typer = "^0.9"
anthropic = "^0.18"
openai = "^1.5"
aiosqlite = "^0.19"
pyyaml = "^6.0"
httpx = "^0.25"
zhipuai = "^2.0"
dashscope = "^1.17"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
pytest-asyncio = "^0.21"
black = "^23.12"
ruff = "^0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dcae = "dcae.cli:app"
